% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.beta.p.r
\name{limits.beta.p}
\alias{limits.beta.p}
\title{Limits for continuos data}
\usage{
limits.beta.p(center, std.dev, sizes, conf, ...)
}
\arguments{
\item{center}{sample/group center statistic.}

\item{std.dev}{within group standard deviation.}

\item{sizes}{sample sizes.}

\item{conf}{a numeric value used t o comput control limits, specifying the confidence level (if 0 < conf < 1)}

\item{...}{further arguments are ignored.}
}
\value{
The function \code{limits.beta.p} returns a matrix with lower and upper control limits.
}
\description{
These function are used to compute the upper and lower control limits.
}
\examples{
limits.beta.p(center = 0.9989597, std.dev = 0.0009362578, conf = 0.9)
}
