\name{restoreSessionAndCreateNewModelFromOSModel}
\alias{restoreSessionAndCreateNewModelFromOSModel}
\title{Function to take a model trained using OS, retore the session and create a new model to be retrained
disease specific survival}
\usage{
  restoreSessionAndCreateNewModelFromOSModel(modelEntity_orig)
}
\description{
  Function to take a model trained using OS, retore the
  session and create a new model to be retrained disease
  specific survival
}
\author{
  Adam Margolin
}

