\name{retrainAllOSModels}
\alias{retrainAllOSModels}
\title{Function to retrain all models originally trained using overall survival and train and evaluate using
disease specific survival}
\usage{
  retrainAllOSModels()
}
\description{
  Function to retrain all models originally trained using
  overall survival and train and evaluate using disease
  specific survival
}
\author{
  Adam Margolin
}

