\name{retrainOSModel}
\alias{retrainOSModel}
\title{Function to retrain one model originally trained using overall survival and train and evaluate using
disease specific survival}
\usage{
  retrainOSModel(modelEntityId_orig, trainingData = NULL)
}
\description{
  Function to retrain one model originally trained using
  overall survival and train and evaluate using disease
  specific survival
}
\author{
  Adam Margolin
}

