% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.beta.p.R
\name{sd.beta.p}
\alias{sd.beta.p}
\title{sd.beta.p}
\usage{
sd.beta.p(data, sizes, std.dev)
}
\arguments{
\item{data}{the observed data values.}

\item{sizes}{sample sizes}

\item{std.dev}{within group standard deviation.}
}
\value{
The function \code{sd.beta.p} returns \code{sd} the standard deviation of the statistic charted.
}
\description{
These functions are used to compute statistics required by the beta chart.
}
\details{
Performs the calculation of the standard deviation \code{sd} that will be used in the construction of the control chart.
}
\examples{
data(Drapper1998data)
sd.beta.p(Drapper1998data)
}
