% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.beta.prop.R
\name{sd.beta.prop}
\alias{sd.beta.prop}
\title{sd.beta.prop}
\usage{
sd.beta.prop(data, sizes, std.dev)
}
\arguments{
\item{data}{the observed data values.}

\item{sizes}{sample sizes}

\item{std.dev}{within group standard deviation.}
}
\description{
These functions are used to compute statistics required by the p chart.
}
\details{
modified qcc functions
}
