% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.beta.prop.r
\name{stats.beta.prop}
\alias{stats.beta.prop}
\title{stats.beta.prop}
\usage{
stats.beta.prop(data,sizes)
}
\arguments{
\item{data}{the observed data values.}

\item{sizes}{sample sizes}
}
\description{
These functions are used to compute statistics required by the p chart.
}
\details{
modified qcc functions
}
