% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.beta.r
\name{stats.beta}
\alias{stats.beta}
\title{stats.beta}
\usage{
stats.beta(data, sizes)
}
\arguments{
\item{data}{the observed data values.}

\item{sizes}{sample sizes}
}
\value{
The function \code{stats.beta} returns a list with components \code{statistics} and \code{center}.
}
\description{
These functions are used to compute statistics required by the beta chart.
}
\details{
Provides a list containing the centerline of the chart \code{pbar} and the data manipulated to be used in the construction of the chart \code{data/sizes}.
}
\examples{
data(Montgomery2005)
stats.beta(Montgomery2005$Defective, Montgomery2005$Sample)
}
