% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.beta.prop.r
\name{limits.beta.prop}
\alias{limits.beta.prop}
\title{limits.beta.prop}
\usage{
limits.beta.prop(center, std.dev, sizes, conf)
}
\arguments{
\item{center, std.dev, conf}{A number}

\item{sizes}{Numeric vector}
}
\description{
modified qcc functions
}
