% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.beta.r
\name{limits.beta}
\alias{limits.beta}
\title{Limits for discrete data}
\usage{
limits.beta(center, std.dev, sizes, conf)
}
\arguments{
\item{center}{sample/group center statistic.}

\item{std.dev}{within group standard deviation.}

\item{sizes}{sample sizes.}

\item{conf}{a numeric value used t o comput control limits, specifying the confidence level (0 < conf < 1)}
}
\value{
The function \code{limits.beta} returns a matrix with lower and upper control limits.
}
\description{
These function are used to compute the upper and lower control limits.
}
\examples{
data(Montgomery2005)
limits.beta(center = 0.2313333, std.dev = 0.421685, sizes = Montgomery2005$Sample, conf = 0.9)
}
