% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.beta.p.R
\name{sd.beta.p}
\alias{sd.beta.p}
\title{sd.beta.p}
\usage{
sd.beta.p(data, sizes, std.dev)
}
\arguments{
\item{data}{the observed data values.}

\item{sizes}{sample sizes}

\item{std.dev}{within group standard deviation.}
}
\description{
These functions are used to compute statistics required by the p chart.
}
\details{
modified qcc functions
}
\examples{
data(montgomery2005)
sd.beta.p(Drapper1998data)
}
