% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.beta.prop.R
\name{sd.beta.prop}
\alias{sd.beta.prop}
\title{sd.beta.prop}
\usage{
sd.beta.prop(data, sizes, std.dev)
}
\arguments{
\item{data, sizes}{Numeric vector}

\item{std.dev}{A number}
}
\description{
modified qcc functions
}
