% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.beta.r
\name{sd.beta}
\alias{sd.beta}
\title{sd.beta}
\usage{
sd.beta(data, sizes, ...)
}
\arguments{
\item{data}{the observed data values.}

\item{sizes}{sample sizes}

\item{...}{further arguments are ignored.}
}
\description{
These functions are used to compute statistics required by the p chart.
}
\details{
modified qcc functions
}
