% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.beta.r
\name{sd.beta}
\alias{sd.beta}
\title{sd.beta}
\usage{
sd.beta(data, sizes, ...)
}
\arguments{
\item{data}{the observed data values.}

\item{sizes}{sample sizes}

\item{...}{further arguments are ignored.}
}
\value{
The function \code{sd.beta} returns \code{std.dev} the standard deviation of the statistic charted.
}
\description{
These functions are used to compute statistics required by the beta chart.
}
\details{
Performs the calculation of the standard deviation \code{std.dev} that will be used in the construction of the control chart.
}
\examples{
data(Montgomery2005)
sd.beta(Montgomery2005$Defective, Montgomery2005$Sample)
}
