% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.beta.r
\name{stats.beta}
\alias{stats.beta}
\title{stats.beta}
\usage{
stats.beta(data, sizes)
}
\arguments{
\item{data}{the observed data values.}

\item{sizes}{sample sizes}
}
\description{
These functions are used to compute statistics required by the p chart.
}
\details{
modified qcc functions
}
\examples{
data(montgomery2005)
stats.beta(Montgomery2005$Defective, Montgomery2005$Sample)
}
